requirements.txt
instagrapi
python-dotenv
Pillow>=8.1.1